requirements.txt
firebase_admin==6.5.0
groq==0.6.0
langchain==0.2.0
langchain_groq==0.1.4
Pillow==10.0.0
python-dotenv==1.0.1
rembg==2.0.56
streamlit==1.34.0
streamlit_option_menu==0.3.12